rename service to mapbox uploader
